+++ node/index.commonjs.js
@@ -1,0 +1,6 @@
+exports = module.exports = require('../commonjs/readXlsxFileNode').default
+exports.Integer = require('../commonjs/types/Integer').default
+exports.Email = require('../commonjs/types/Email').default
+exports.URL = require('../commonjs/types/URL').default
+exports.parseExcelDate = require('../commonjs/parseDate').default
+exports['default'] = require('../commonjs/readXlsxFileNode').default