+++ node/index.js
@@ -1,0 +1,5 @@
+export { default as default } from '../modules/readXlsxFileNode'
+export { default as Integer } from '../modules/types/Integer'
+export { default as Email } from '../modules/types/Email'
+export { default as URL } from '../modules/types/URL'
+export { default as parseExcelDate } from '../modules/parseDate'