+++ schema/index.commonjs.js
@@ -1,0 +1,2 @@
+exports = module.exports = require('../commonjs/convertToJson').default
+exports['default'] = require('../commonjs/convertToJson').default