+++ source/readXlsxFileContents.js
@@ -1,10 +1,14 @@
 import readXlsx from './readXlsx'
 import convertToJson from './convertToJson'
+import convertMapToSchema from './convertMapToSchema'
 
-export default function readXlsxFileContents(entries, xml, options) {
-	const result = readXlsx(entries, xml, { ...options, properties: options.schema || options.properties })
-	if (options.schema) {
-		return convertToJson(result.data, options.schema, { ...options, properties: result.properties })
+export default function readXlsxFileContents(entries, xml, { schema, map, ...options}) {
+	if (!schema && map) {
+		schema = convertMapToSchema(map)
+	}
+	const result = readXlsx(entries, xml, { ...options, properties: schema || options.properties })
+	if (schema) {
+		return convertToJson(result.data, schema, { ...options, properties: result.properties })
 	}
 	return result
 }